pyproject.toml
[project]
name = "docling-rag-agent"
version = "0.1.0"
description = "Streamlit-based RAG agent with PostgreSQL/PGVector"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic-ai>=0.7.4",
    "asyncpg>=0.30.0",
    "numpy>=2.0.2",
    "openai>=1.0.0",
    "docling[vlm]>=2.55.0",
    "streamlit>=1.31.0",
    "nest-asyncio>=1.6.0",
    "watchdog>=4.0.0",
    "fastmcp>=0.1.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.30.1",
    "langfuse>=3.0.0",
    "prometheus_client>=0.19.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
# E501: line too long (handled by formatter)
# E402: module level import not at top (common in Streamlit)
# W291/W293: trailing/blank line whitespace (common in embedded SQL)
# F841: local variable assigned but never used (common in tests for context managers)
ignore = ["E501", "E402", "W291", "W293", "F841"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true

[tool.coverage.run]
source = ["core", "ingestion", "docling_mcp", "utils"]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/conftest.py",
    "*/test_*.py"
]

[tool.coverage.report]
fail_under = 70
precision = 2
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod"
]

[tool.coverage.html]
directory = "htmlcov"
